Fix #22, encode error in python2
